Move LocationInfoView to Location model and updated ViewModel to reference new class
